Move query generation to separate function
