+++ src/client/visualizations/base-visualization/base-visualization.tsx
@@ -15,19 +15,16 @@
  * limitations under the License.
  */
 
-import { $, Dataset, Expression, LimitExpression, ply, SortExpression } from "plywood";
+import { Dataset, Expression } from "plywood";
 import * as React from "react";
 import { Essence } from "../../../common/models/essence/essence";
-import { toExpression as filterClauseToExpression } from "../../../common/models/filter-clause/filter-clause";
-import { CurrentFilter, Measure, MeasureDerivation, PreviousFilter } from "../../../common/models/measure/measure";
-import { toExpression as splitToExpression } from "../../../common/models/split/split";
+import { Measure } from "../../../common/models/measure/measure";
 import { Timekeeper } from "../../../common/models/timekeeper/timekeeper";
 import { DatasetLoad, VisualizationProps } from "../../../common/models/visualization-props/visualization-props";
-import { sortDirectionMapper } from "../../../common/view-definitions/version-3/split-definition";
+import makeQuery from "../../../common/utils/query/visualization-query";
 import { GlobalEventListener } from "../../components/global-event-listener/global-event-listener";
 import { Loader } from "../../components/loader/loader";
 import { QueryError } from "../../components/query-error/query-error";
-import { SPLIT } from "../../config/constants";
 import "./base-visualization.scss";
 
 export interface BaseVisualizationState {
@@ -73,99 +70,7 @@ export class BaseVisualization<S extends BaseVisualizationState> extends React.C
   }
 
   protected makeQuery(essence: Essence, timekeeper: Timekeeper): Expression {
-    const { splits, colors, dataCube } = essence;
-    const measures = essence.getEffectiveMeasures();
-
-    const $main = $("main");
-
-    const hasComparison = essence.hasComparison();
-    const mainFilter = essence.getEffectiveFilter(timekeeper, { combineWithPrevious: hasComparison, highlightId: this.id });
-
-    const currentFilter = filterClauseToExpression(essence.currentTimeFilter(timekeeper));
-    const previousFilter = hasComparison ? filterClauseToExpression(essence.previousTimeFilter(timekeeper)) : null;
-
-    const mainExp: Expression = ply().apply("main", $main.filter(mainFilter.toExpression()));
-
-    function applyMeasures(query: Expression, nestingLevel = 0): Expression {
-      return measures.reduce((query, measure) => {
-        if (!hasComparison) {
-          return query.performAction(
-            measure.toApplyExpression(nestingLevel)
-          );
-        }
-        return query
-          .performAction(measure.toApplyExpression(nestingLevel, new CurrentFilter(currentFilter)))
-          .performAction(measure.toApplyExpression(nestingLevel, new PreviousFilter(previousFilter)));
-      }, query);
-    }
-
-    const queryWithMeasures = applyMeasures(mainExp);
-
-    function applySplit(i: number): Expression {
-      const split = splits.getSplit(i);
-      const splitDimension = dataCube.getDimension(split.reference);
-      const { sort, limit } = split;
-      if (!sort) {
-        throw new Error("something went wrong during query generation");
-      }
-
-      const currentSplit = splitToExpression(split, hasComparison && currentFilter, hasComparison && essence.timeShift.valueOf());
-      let subQuery: Expression =
-        $main.split(currentSplit, splitDimension.name);
-
-      if (colors && colors.dimension === splitDimension.name) {
-        const havingFilter = colors.toHavingFilter(splitDimension.name);
-        if (havingFilter) {
-          subQuery = subQuery.performAction(havingFilter);
-        }
-      }
-
-      const nestingLevel = i + 1;
-
-      subQuery = applyMeasures(subQuery, nestingLevel);
-
-      // It's possible to define sort on measure that's not selected thus we need to add apply expression for that measure.
-      // We don't need add apply expressions for:
-      //   * dimensions - they're already defined as apply expressions because of splits
-      //   * selected measures - they're defined as apply expressions already
-      //   * previous - we need to define them earlier so they're present here
-      const { name: sortMeasureName, derivation } = Measure.nominalName(sort.reference);
-      if (sortMeasureName && derivation === MeasureDerivation.CURRENT) {
-        const sortMeasure = dataCube.getMeasure(sortMeasureName);
-        if (sortMeasure && !measures.contains(sortMeasure)) {
-          subQuery = subQuery.performAction(sortMeasure.toApplyExpression(nestingLevel, new CurrentFilter(currentFilter)));
-        }
-      }
-      if (sortMeasureName && derivation === MeasureDerivation.DELTA) {
-        subQuery = subQuery.apply(sort.reference, $(sortMeasureName).subtract($(Measure.derivedName(sortMeasureName, MeasureDerivation.PREVIOUS))));
-      }
-      subQuery = subQuery.performAction(new SortExpression({
-        expression: $(sort.reference),
-        direction: sortDirectionMapper[sort.direction]
-      }));
-
-      if (colors && colors.dimension === splitDimension.name) {
-        subQuery = subQuery.performAction(colors.toLimitExpression());
-      } else if (limit) {
-        subQuery = subQuery.performAction(new LimitExpression({ value: limit }));
-      } else if (splitDimension.kind === "number") {
-        // Hack: Plywood converts groupBys to topN if the limit is below a certain threshold.  Currently sorting on dimension in a groupBy query does not
-        // behave as expected and in the future plywood will handle this, but for now add a limit so a topN query is performed.
-        // 5000 is just a randomly selected number that's high enough that it's not immediately obvious that there's a limit.
-        subQuery = subQuery.limit(5000);
-      }
-
-      if (i + 1 < splits.length()) {
-        subQuery = subQuery.apply(SPLIT, applySplit(i + 1));
-      }
-
-      return subQuery;
-    }
-
-    if (splits.length() > 0) {
-      return queryWithMeasures.apply(SPLIT, applySplit(0));
-    }
-    return queryWithMeasures;
+    return makeQuery(essence, timekeeper);
   }
 
   protected fetchData(essence: Essence, timekeeper: Timekeeper): void {
